requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
crispy-bootstrap4==2024.1
Django==5.0.3
django-bootstrap4==24.1
django-crispy-forms==2.1
django-filter==24.2
djangorestframework==3.15.1
Markdown==3.6
python-dotenv==1.0.1
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
